Reduce `recv` calls by 1024x for XML and JSON clients.
Functionally identical, but instead of reading in byte by byte we we can
read a chunk that is reasonably sized for standard TCP packets. Chunk size can
be tweaked but 1024 bytes is a very common value and the return on further
optimization has quickly diminishing returns.

This reduces several requests from ~6-10k calls to `recv` to 6-10.